+++ apps/web/src/config/constants/swapWarningTokens.ts
@@ -1,12 +1,15 @@
 import { ChainId } from '@pancakeswap/chains'
 import { Token } from '@pancakeswap/sdk'
 import { bscTokens, ethereumTokens, zksyncTokens } from '@pancakeswap/tokens'
-import { bscWarningTokens } from 'config/constants/warningTokens'
+import { baseWarningTokens } from 'config/constants/warningTokens/baseWarningTokens'
+import { bscWarningTokens } from 'config/constants/warningTokens/bscWarningTokens'
 
 const { alETH } = ethereumTokens
 const { bondly, itam, ccar, bttold, abnbc, metis } = bscTokens
 const { pokemoney, free, safemoon, gala, xcad, lusd, nfp } = bscWarningTokens
 const { usdPlus } = zksyncTokens
+const { ath } = baseWarningTokens
+
 interface WarningTokenList {
   [chainId: number]: {
     [key: string]: Token
@@ -34,6 +37,9 @@ const SwapWarningTokens = <WarningTokenList>{
   },
   [ChainId.ZKSYNC]: {
     usdPlus,
+  },
+  [ChainId.BASE]: {
+    ath,
   },
 }