+++ packages/chains/src/utils.ts
@@ -1,5 +1,11 @@
 import { ChainId, testnetChainIds } from './chainId'
-import { chainNameToChainId, chainNames, chainNamesInKebabCase, defiLlamaChainNames } from './chainNames'
+import {
+  chainNameToChainId,
+  chainNames,
+  chainNamesInKebabCase,
+  defiLlamaChainNames,
+  mainnetChainNamesInKebabCase,
+} from './chainNames'
 
 export function getChainName(chainId: ChainId) {
   return chainNames[chainId]
@@ -7,6 +13,10 @@ export function getChainName(chainId: ChainId) {
 
 export function getChainNameInKebabCase(chainId: ChainId) {
   return chainNamesInKebabCase[chainId]
+}
+
+export function getMainnetChainNameInKebabCase(chainId: keyof typeof mainnetChainNamesInKebabCase) {
+  return mainnetChainNamesInKebabCase[chainId]
 }
 
 export function getLlamaChainName(chainId: ChainId) {