+++ packages/universal-router-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pancakeswap/universal-router-sdk
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies [9a16780]
+  - @pancakeswap/chains@0.4.6
+  - @pancakeswap/permit2-sdk@1.0.9
+  - @pancakeswap/smart-router@6.0.13
+  - @pancakeswap/sdk@5.8.7
+  - @pancakeswap/v3-sdk@3.8.8
+
 ## 1.0.25
 
 ### Patch Changes